requirements.txt
Flask_JWT_Extended==3.14.0
six==1.11.0
scipy==1.1.0
pandas==0.23.4
schedule==0.5.0
pymongo==3.7.0
Flask_RESTful==0.3.6
setuptools==39.1.0
requests==2.18.4
Flask_Cors==3.0.4
numpy==1.15.4
Keras==2.2.2
Flask==1.0.2
Flask_MonitoringDashboard==2.0.6
alpha_vantage==2.1.0
python_crontab==2.3.6
python_dateutil==2.8.0
mailer==0.8.1
scikit_learn==0.20.2
